Move Traits dict into its own module.
